fix(shared): Fix CI compilation errors and implement stubbed functions
Fix compilation errors causing Main Branch Quality workflow to fail.
Also implemented previously stubbed autograd functions.

Key fixes:
- Fix scalar ExTensor constructors to fully initialize all fields
  instead of using invalid delegating constructor pattern
- Fix partial struct destruction errors in conv.mojo and config.mojo
  by copying fields before transfer
- Add explicit __copyinit__ for SavedTensors and TapeNode in tape.mojo
- Fix Dict iteration pattern in arg_parser.mojo
- Add raises keyword to functions calling raising functions
- Fix test files: DType comparison, math imports, SIMD types

Implemented functions:
- SGD.step(): Updates parameters using gradients from tape
- SGD.zero_grad(): Clears gradient registry
- GradientTape.backward(): Full backward pass implementation
  with index-based node access to avoid TapeNode copying

Note: _hex_to_bytes in io.mojo validates input but cannot write to
output due to UnsafePointer origin constraints in Mojo v0.25.7.

Closes #2280

+++ tests/core/types/test_mxfp4_block.mojo
@@ -12,6 +12,7 @@ All tests use pure functional API.
 """
 
 from math import isinf, isnan
+
 from shared.core.types.mxfp4 import MXFP4, MXFP4Block, E8M0Scale
 from shared.core.types.fp4 import FP4_E2M1
 from tests.shared.conftest import (
@@ -138,8 +139,8 @@ fn test_mxfp4_block_roundtrip_mixed_signs() raises:
     # Verify signs are preserved
     for i in range(32):
         var expected = (Float32(1.0) if i % 2 == 0 else Float32(-1.0)) * Float32(i) * Float32(0.1)
-        var expected_sign = 1.0 if expected >= 0 else -1.0
-        var decoded_sign = 1.0 if decoded[i] >= 0 else -1.0
+        var expected_sign = Float32(1.0) if expected >= 0 else Float32(-1.0)
+        var decoded_sign = Float32(1.0) if decoded[i] >= 0 else Float32(-1.0)
         assert_equal(Int(expected_sign), Int(decoded_sign))
